Reorganize files and functions
